+++ src/solidityByExample/function.sol
@@ -2,58 +2,48 @@
 pragma solidity ^0.8.24;
 
 contract Function {
-    function returnMany() public pure returns (uint, bool, uint) {
+    function returnMany() public pure returns (uint256, bool, uint256) {
         return (1, true, 2);
     }
 
-    function named() public pure returns (uint x, bool b, uint y) {
+    function named() public pure returns (uint256 x, bool b, uint256 y) {
         return (1, true, 2);
     }
 
-    function assigned() public pure returns (uint x, bool b, uint y) {
-        x = 1; 
+    function assigned() public pure returns (uint256 x, bool b, uint256 y) {
+        x = 1;
         b = true;
         y = 2;
     }
 
-    function destructuringnments() public pure returns (uint, bool, uint, uint, uint) {
-        (uint i, bool b, uint j) = returnMany();
-        (uint x,, uint y) = (4, 5, 6);
+    function destructuringnments() public pure returns (uint256, bool, uint256, uint256, uint256) {
+        (uint256 i, bool b, uint256 j) = returnMany();
+        (uint256 x,, uint256 y) = (4, 5, 6);
 
         return (i, b, j, x, y);
     }
 
-    function arrayInput(uint[] memory _arr) public {}
+    function arrayInput(uint256[] memory _arr) public {}
 
-    uint[] public arr;
+    uint256[] public arr;
 
-    function arrayOutput() public view returns (uint[] memory) {
+    function arrayOutput() public view returns (uint256[] memory) {
         return arr;
     }
 }
 
 contract XYZ {
-    function someFuncWithManyInputs(
-        uint x, 
-        uint y,
-        uint z,
-        address a,
-        bool b,
-        string memory c
-    ) public pure returns (uint) {}
+    function someFuncWithManyInputs(uint256 x, uint256 y, uint256 z, address a, bool b, string memory c)
+        public
+        pure
+        returns (uint256)
+    {}
 
-    function callFunc() external pure returns (uint) {
+    function callFunc() external pure returns (uint256) {
         return someFuncWithManyInputs(1, 2, 3, address(0), true, "c");
     }
 
-    function callFuncWithKeyValue() external pure returns (uint) {
-        return someFuncWithManyInputs({
-            a: address(0),
-            b: true, 
-            c: "c",
-            x: 1,
-            y: 2,
-            z: 3
-        });
+    function callFuncWithKeyValue() external pure returns (uint256) {
+        return someFuncWithManyInputs({a: address(0), b: true, c: "c", x: 1, y: 2, z: 3});
     }
 }
