+++ src/solidityByExample/structs.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.24;
 
 contract Todos {
-    struct Todo{
+    struct Todo {
         string text;
         bool completed;
     }
@@ -11,7 +11,7 @@ contract Todos {
 
     function create(string calldata _text) public {
         todos.push(Todo(_text, false));
-        todos.push(Todo({text: _text,completed: false}));
+        todos.push(Todo({text: _text, completed: false}));
 
         Todo memory todo;
         todo.text = _text;
@@ -19,17 +19,17 @@ contract Todos {
         todos.push(todo);
     }
 
-    function get(uint _index) public view returns (string memory text, bool completed) {
+    function get(uint256 _index) public view returns (string memory text, bool completed) {
         Todo storage todo = todos[_index];
         return (todo.text, todo.completed);
     }
 
-    function updateText(uint _index, string calldata _text) public {
+    function updateText(uint256 _index, string calldata _text) public {
         Todo storage todo = todos[_index];
         todo.text = _text;
     }
 
-    function toggleCompleted(uint _index) public {
+    function toggleCompleted(uint256 _index) public {
         Todo storage todo = todos[_index];
         todo.completed = !todo.completed;
     }
